Justeringer etter qa
